requirements.txt
streamlit
psycopg[binary]
python-dotenv
